Fix pace precision loss for imperial unit splits

+++ api/Endpoints/WorkoutsEndpoints.cs
@@ -28,10 +28,10 @@ public static class WorkoutsEndpoints
         public DateTime StartedAt { get; set; }
         public int DurationS { get; set; }
         public double DistanceM { get; set; }
-        public int AvgPaceS { get; set; }
+        public double AvgPaceS { get; set; }
         public double? SimilarityScore { get; set; }
         public int? TimeDifferenceS { get; set; } // Negative = faster, Positive = slower
-        public int? PaceDifferenceS { get; set; } // Negative = faster pace
+        public double? PaceDifferenceS { get; set; } // Negative = faster pace
         public int? RelativeEffort { get; set; }
         public double? ElevGainM { get; set; }
     }
@@ -2657,7 +2657,7 @@ public static class WorkoutsEndpoints
         DateTime startedAtUtc,
         int durationSeconds,
         double distanceMeters,
-        int avgPaceS,
+        double avgPaceS,
         double? elevationGainMeters,
         byte[] rawFileData,
         string fileName,
@@ -3161,7 +3161,7 @@ public static class WorkoutsEndpoints
 
             // Calculate average pace (seconds per km - stored in metric)
             var avgPaceS = distanceMeters > 0 && durationSeconds > 0
-                ? (int)(durationSeconds / (distanceMeters / 1000.0))
+                ? durationSeconds / (distanceMeters / 1000.0)
                 : 0;
 
             // Extract additional metrics from RawGpxData JSON (for GPX files)
